Compl. initial work on .md -> pdf (via LaTeX)
This is the initial work on the markdown to .pdf, html, and epub static
site generator. In this commit, we create the directory structure for
the program, which takes the form of a pandoc folder, along with source
files and output targets.

Right now, only the latex output is working. In future commits, epub and
html outputs will be made as well.