Refactored as a command line program
Originally, the polytexy program was called static-site-gen, which was
intended to be a static site generator. The scale and scope of a static
site generator was too large for my current needs, so I decided to
embark on a major refactoring of the program and make it a command line
utility instead.

Now renamed as polytexy, the program is designed to take a markdown file
as input, and output a pdf, html, and epub file. Polytexy will be
refined, and eventually a component of the larger static site generator.