Refactored polytexy.sh
In this major refactoring of polytexy.sh, I added basic sanity checks
for dependencies, better usage instructions, as well as move the build
commands into functions. This ensures that no commands will be run
without at least ensuring that the dependencies are met, and that the
input files are actually markdown files.