Updated README.md w/ installation, usage, info
This commit adds proper documentation to `polytexy` for the first time,
giving installation details and some information on usage and features.
Right now, the project is still not complete, and neither is the
documentation - but this should serve as a good foundation for further
info.

A vector logo is also provided for the project.