Add customisation options for `\maketitle` macro
The default `\documentclass{article}` `\maketitle` macro has the option
of including the author and the date. This commit uses the `titling`
package to allow some ability to customise the creation of document
titles, by allowing us to optionally omit the date and/or the author.
When the date and author are specified, but `omitdate` and/or
`omitauthor` is set to true, the date and author will not be typeset in
the document, even though this information will be included as PDF
metadata.

Additionally, in order to better support language locales, we moved to
using the `polyglossia` package instead of `babel`, and also removed the
declaration of the language type from the documentclass options.
Instead, the document language is specified from the `lang` YAML
attribute, and that is applied both for PDF/X metadata as well as for
the document itself.